added Typed Redux Action Creators from `ts-redux-actions`
