+++ src/mapped-types.ts
@@ -16,18 +16,21 @@ export type Diff<T extends string, U extends string> = (
   & { [x: string]: never }
 )[T];
 // TestDiff expects: 'a' | 'b'
-type TestDiff = Diff<'a' | 'b' | 'c', 'c' | 'd'>;
+type TestDiff =
+  Diff<'a' | 'b' | 'c', 'c' | 'd'>;
 
 /** Omit */
 export type Omit<T, K extends keyof T> = {
   [P in Diff<keyof T, K>]: T[P]
 };
 // TestOmit expects: { b: number, c: boolean }
-type TestOmit = Omit<{ a: string, b: number, c: boolean }, 'a'>;
+type TestOmit =
+  Omit<{ a: string, b: number, c: boolean }, 'a'>;
 
 /** Overwrite */
 export type Overwrite<T, U> = {
   [P in Diff<keyof T, keyof U>]: T[P]
 } & U;
 // TestOverwrite expects: { b: number, c: boolean } & { a: number }
-type TestOverwrite = Overwrite<{ a: string, b: number, c: boolean }, { a: number }>;
+type TestOverwrite =
+  Overwrite<{ a: string, b: number, c: boolean }, { a: number }>;
