+++ src/v2/create-action-creator.ts
@@ -1,3 +1,5 @@
+// DEPRECATED: delete soon after new Major is realeased
+
 export type EmptyActionCreator<T extends string> = {
   (): { type: T }, type?: T,
 };
@@ -11,9 +13,23 @@ export type PayloadSelector<S, M, P> = (state: S, meta?: M) => P;
 export function createActionCreator<T extends string, S, M, P>(
   type: T, payloadSelector: PayloadSelector<S, M, P>,
 ): PayloadActionCreator<T, S, M, P>;
+
 export function createActionCreator<T extends string>(
   type: T,
 ): EmptyActionCreator<T>;
+
+/**
+ *
+ * @deprecated
+ * @export
+ * @template T
+ * @template S
+ * @template M
+ * @template P
+ * @param {T} type
+ * @param {PayloadSelector<S, M, P>} [payloadSelector]
+ * @returns {(PayloadActionCreator<T, S, M, P> | EmptyActionCreator<T>)}
+ */
 export function createActionCreator<T extends string, S, M, P>(
   type: T, payloadSelector?: PayloadSelector<S, M, P>,
 ): PayloadActionCreator<T, S, M, P> | EmptyActionCreator<T> {
