+++ src/v2/helpers-test.ts
@@ -1,3 +1,8 @@
+// API v2 - DEPRECATED
+import { EmptyAction, PayloadAction } from '..';
+
+console.log('\nRunning test suite...\n');
+
 // Test Helpers
 const passed = (describe: string) => {
   console.log(describe, true);
@@ -8,8 +13,8 @@ const failed = (describe: string, state: object) => {
 
 // Testing createEmptyAction, createPayloadAction
 import {
-  EmptyAction, PayloadAction, createEmptyAction, createPayloadAction,
-} from '.';
+  createEmptyAction, createPayloadAction,
+} from './helpers';
 
 export namespace TestHelpers1 {
   const APP_STARTED = 'app/APP_STARTED';
@@ -81,7 +86,7 @@ export namespace TestHelpers1 {
 // Testing ActionCreator
 import {
   ActionCreator,
-} from '.';
+} from '..';
 
 export namespace TestHelpers2 {
   export const ActionCreators = {
