--- tests.ts
@@ -1,7 +1,0 @@
-console.log('\nRunning test suite...\n');
-
-import './src/redux-utils.spec';
-import './src/type-utils.spec';
-
-// deprecated
-import './src/helpers.spec';
